Support org-adapt-indentation set to t
In case `org-adapt-indentation` is set to `t` the text of an org node is indented according to the depth of the headline of that node.

The regex checking if we are on a source code block line currently assumes that the start of the code block is always at the start of the line, not taking into account that a user might have `org-adapt-indentation` enabled. Allowing optional spaces before the `#begin` allows completions via `corg` in such setups.